Move TracyCallstack (and dependencies) to common/ folder
