+++ CMakeLists.txt
@@ -177,17 +177,12 @@ set(client_includes
     ${TRACY_PUBLIC_DIR}/client/tracy_SPSCQueue.h
     ${TRACY_PUBLIC_DIR}/client/TracyKCore.hpp
     ${TRACY_PUBLIC_DIR}/client/TracyArmCpuTable.hpp
-    ${TRACY_PUBLIC_DIR}/client/TracyCallstack.h
-    ${TRACY_PUBLIC_DIR}/client/TracyCallstack.hpp
     ${TRACY_PUBLIC_DIR}/client/TracyCpuid.hpp
-    ${TRACY_PUBLIC_DIR}/client/TracyDebug.hpp
     ${TRACY_PUBLIC_DIR}/client/TracyDxt1.hpp
-    ${TRACY_PUBLIC_DIR}/client/TracyFastVector.hpp
     ${TRACY_PUBLIC_DIR}/client/TracyLock.hpp
     ${TRACY_PUBLIC_DIR}/client/TracyProfiler.hpp
     ${TRACY_PUBLIC_DIR}/client/TracyRingBuffer.hpp
     ${TRACY_PUBLIC_DIR}/client/TracyScoped.hpp
-    ${TRACY_PUBLIC_DIR}/client/TracyStringHelpers.hpp
     ${TRACY_PUBLIC_DIR}/client/TracySysPower.hpp
     ${TRACY_PUBLIC_DIR}/client/TracySysTime.hpp
     ${TRACY_PUBLIC_DIR}/client/TracySysTrace.hpp
@@ -199,13 +194,18 @@ set(common_includes
     ${TRACY_PUBLIC_DIR}/common/TracyAlign.hpp
     ${TRACY_PUBLIC_DIR}/common/TracyAlloc.hpp
     ${TRACY_PUBLIC_DIR}/common/TracyApi.h
+    ${TRACY_PUBLIC_DIR}/common/TracyCallstack.h
+    ${TRACY_PUBLIC_DIR}/common/TracyCallstack.hpp
     ${TRACY_PUBLIC_DIR}/common/TracyColor.hpp
+    ${TRACY_PUBLIC_DIR}/common/TracyDebug.hpp
+    ${TRACY_PUBLIC_DIR}/common/TracyFastVector.hpp
     ${TRACY_PUBLIC_DIR}/common/TracyForceInline.hpp
     ${TRACY_PUBLIC_DIR}/common/TracyMutex.hpp
     ${TRACY_PUBLIC_DIR}/common/TracyProtocol.hpp
     ${TRACY_PUBLIC_DIR}/common/TracyQueue.hpp
     ${TRACY_PUBLIC_DIR}/common/TracySocket.hpp
     ${TRACY_PUBLIC_DIR}/common/TracyStackFrames.hpp
+    ${TRACY_PUBLIC_DIR}/common/TracyStringHelpers.hpp
     ${TRACY_PUBLIC_DIR}/common/TracySystem.hpp
     ${TRACY_PUBLIC_DIR}/common/TracyUwp.hpp
     ${TRACY_PUBLIC_DIR}/common/TracyYield.hpp)
