+++ cmake/server.cmake
@@ -3,6 +3,7 @@ set(TRACY_COMMON_DIR ${CMAKE_CURRENT_LIST_DIR}/../public/common)
 set(TRACY_COMMON_SOURCES
     tracy_lz4.cpp
     tracy_lz4hc.cpp
+    TracyCallstack.cpp
     TracySocket.cpp
     TracyStackFrames.cpp
     TracySystem.cpp
