+++ meson.build
@@ -145,17 +145,12 @@ client_includes = [
     'public/client/tracy_rpmalloc.hpp',
     'public/client/tracy_SPSCQueue.h',
     'public/client/TracyArmCpuTable.hpp',
-    'public/client/TracyCallstack.h',
-    'public/client/TracyCallstack.hpp',
-    'public/client/TracyDebug.hpp',
     'public/client/TracyDxt1.hpp',
-    'public/client/TracyFastVector.hpp',
     'public/client/TracyKCore.hpp',
     'public/client/TracyLock.hpp',
     'public/client/TracyProfiler.hpp',
     'public/client/TracyRingBuffer.hpp',
     'public/client/TracyScoped.hpp',
-    'public/client/TracyStringHelpers.hpp',
     'public/client/TracySysPower.hpp',
     'public/client/TracySysTime.hpp',
     'public/client/TracySysTrace.hpp',
@@ -168,13 +163,18 @@ common_includes = [
     'public/common/TracyAlign.hpp',
     'public/common/TracyAlloc.hpp',
     'public/common/TracyApi.h',
+    'public/common/TracyCallstack.h',
+    'public/common/TracyCallstack.hpp',
     'public/common/TracyColor.hpp',
+    'public/common/TracyDebug.hpp',
+    'public/common/TracyFastVector.hpp',
     'public/common/TracyForceInline.hpp',
     'public/common/TracyMutex.hpp',
     'public/common/TracyProtocol.hpp',
     'public/common/TracyQueue.hpp',
     'public/common/TracySocket.hpp',
     'public/common/TracyStackFrames.hpp',
+    'public/common/TracyStringHelpers.hpp',
     'public/common/TracySystem.hpp',
     'public/common/TracyUwp.hpp',
     'public/common/TracyYield.hpp'
