+++ public/client/TracyProfiler.cpp
@@ -71,8 +71,8 @@
 #include "../common/TracyYield.hpp"
 #include "../common/tracy_lz4.hpp"
 #include "tracy_rpmalloc.hpp"
-#include "TracyCallstack.hpp"
-#include "TracyDebug.hpp"
+#include "../common/TracyCallstack.hpp"
+#include "../common/TracyDebug.hpp"
 #include "TracyDxt1.hpp"
 #include "TracyScoped.hpp"
 #include "TracyProfiler.hpp"
