+++ public/common/TracyStringHelpers.hpp
